pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "extractmc"
version = "1.0.0"
description = "MasterCard IPM file parser — extracts records to CSV / JSON"
requires-python = ">=3.11"
dependencies = [
    "hexdump>=3.3",
    "PyYAML>=6.0",
    "typer>=0.9",
]

[project.scripts]
mci-extract = "mci.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
mci = ["mideu.yml"]